fix 1v1 being stuck after a revive or buy phase & refactor isOneVersusOne into a own service so there aren't so many duplicates
